Jam SqlBuffer.netfx components into SqlBuffer class in common project
